Hook up CLPs to the swap page. Refactored Store into a bunch of smaller files. Still need to implement the actual swap execution.
