+++ gui/src/store/Price.ts
@@ -1,0 +1,8 @@
+import { Instance, types } from 'mobx-state-tree'
+
+export const Price = types.model({
+  price: types.number,
+  symbol: types.string,
+})
+
+export interface IPrice extends Instance<typeof Price> {}
